+++ tests/testthat/test-DBItest.R
@@ -2,8 +2,8 @@ context("DBI tests")
 
 DBItest::make_context(SQLServer(), list(server = "TEST", database = "DBITest"))
 DBItest::test_getting_started()
-DBItest::test_driver(skip = c("data_type", "stress_load_unload"))
-DBItest::test_connection(skip = "stress_load_connect_unload")
+DBItest::test_driver()
+DBItest::test_connection()
 # DBItest::test_result(skip = c(
 #   # jTDS closes open statements when closing connections. Closing statements
 #   # closes corresponding resultsets (per JDBC doc). So no need to test for this.